Login details

+++ src/main/java/pl/coderslab/charity/Services/DonationService.java
@@ -5,7 +5,7 @@ import pl.coderslab.charity.Models.Donation;
 public interface DonationService {
 
     void save(Donation donation);
-    int getQuantityDonations();
+    Long getQuantityDonations();
 
 
 }
